requirements.txt
numpy>=2.3.0
pandas>=2.3.0
scikit-learn>=1.7.0
matplotlib>=3.10.3
seaborn>=0.12.0
tqdm>=4.67.1
faiss-cpu>=1.11.0
notebook>=7.4.3